feat: 모달 외부를 클릭하여 모달 닫기 기능 추가
